+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
 ## v1.2
+- **Added scroll bars to the text fields**
 - **Auto-Saves every 60 seconds so you never lose any work**
 - **Can use both `Return` and `Enter` (`Opt-Return`) to add newlines, as opposed to _just_ `Enter`**
 - **Allows user to change font of message and response text fields to accommodate needs**
 - **Changed icon to something a little more exciting**
+- Added a "Cancel" button to the popup that adds a new recipient
 - Fixed issue when changing recipients where it doesn't repopulate the NSTextFields
 - Fixed an issue where font is system default
 - Now stores font size and name preferences on save so that a preference doesn't have to be changed every time if it's not the default